+++ src/autoscaler/api/binding_request/binding-configuration.json
