+++ src/autoscaler/api/binding_request/binding-request-parser_suite_test.go
@@ -1,4 +1,4 @@
-package binding_requests_test
+package binding_request_test
 
 import (
 	. "github.com/onsi/ginkgo/v2"
