+++ src/autoscaler/api/binding_request/binding-request.json
