+++ src/autoscaler/api/binding_request/combined_parser/parser.go
@@ -6,7 +6,7 @@ import (
 	"github.com/xeipuuv/gojsonschema"
 
 	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/models"
-	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/api/binding_requests"
+	"code.cloudfoundry.org/app-autoscaler/src/autoscaler/api/binding_request"
 )
 
 // Combined parser that tries out all other parser that are associated to it in order
@@ -14,11 +14,11 @@ import (
 //
 // In case all associated parser fail, it returns the error of the first one.
 type CombinedBindingRequestParser struct {
-	parsers []binding_requests.Parser
+	parsers []binding_request.Parser
 }
 
 func (p CombinedBindingRequestParser) Parse(
 	bindingReqParams string,
-) (binding_requests.Parameters, error) {
-	return binding_requests.Parameters{}, models.ErrUnimplemented // 🚧 To-do
+) (binding_request.Parameters, error) {
+	return binding_request.Parameters{}, models.ErrUnimplemented // 🚧 To-do
 }
