+++ src/autoscaler/api/binding_request/shared_definitions.json
